dom/indexedDB/test/test_objectCursors.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const name = window.location.pathname;
      const description = "My Test Database";

      const objectStores = [
        { name: "a", autoIncrement: false },
        { name: "b", autoIncrement: true }
      ];

      const indexes = [
        { name: "a", options: { } },
        { name: "b", options: { unique: true } }
      ];

      let request = mozIndexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.target.result;

      for (let i in objectStores) {
        request = db.setVersion("1");
        request.onerror = errorHandler;
        request.onsuccess = grabEventAndContinueHandler;
        let event = yield;

        let objectStore =
          db.createObjectStore(objectStores[i].name,
                               { keyPath: "id",
                                 autoIncrement: objectStores[i].autoIncrement });

        for (let j in indexes) {
          objectStore.createIndex(indexes[j].name, "name", indexes[j].options);
        }

        let data = { name: "Ben" };
        if (!objectStores[i].autoIncrement) {
          data.id = 1;
        }

        request = objectStore.add(data);
        request.onerror = errorHandler;
        request.onsuccess = grabEventAndContinueHandler;
        event = yield;

        ok(event.target.result == 1 || event.target.result == 2, "Good id");
      }

      SimpleTest.executeSoon(function() { testGenerator.next(); });
      yield;

      let objectStore = event.target.result;

      for (let i in objectStores) {
        for (let j in indexes) {
          let objectStore = db.transaction(objectStores[i].name)
                              .objectStore(objectStores[i].name);
          let index = objectStore.index(indexes[j].name);

          request = index.openCursor();
          request.onerror = errorHandler;
          request.onsuccess = function (event) {
            is(event.target.result.value.name, "Ben", "Good object");
            SimpleTest.executeSoon(function() { testGenerator.next(); });
          }
          yield;
        }
      }

      finishTest();
      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
